requirements.txt
numpy
scipy
scikit-learn
seaborn
tqdm
paho-mqtt